move objects into objects folder
