+++ apps/docs/src/examples/core/three-arcade-game/game/objects/Ball/Ball.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
-  import { game } from '../Game.svelte'
+  import { game } from '../../Game.svelte'
   import BallOut from './BallOut.svelte'
   import DynamicBall from './DynamicBall.svelte'
   import StaticBall from './StaticBall.svelte'
 </script>
 
 {#if game.state === 'playing'}
-  <DynamicBall />
+  <DynamicBall startAtPosX={game.playerPosition} />
 {:else if game.state === 'game-over'}
   <BallOut />
 {:else}
