+++ apps/docs/src/examples/core/three-arcade-game/game/objects/CustomGridHelper.ts
